Create separate test files for Windows and Linux/MacOS
