+++ _generated/CopyFilesOverSSHV0/Tests/L0.ts
@@ -1,11 +1,11 @@
 import * as utilsTests from './L0UtilsTests';
 
 describe('CopyFilesOverSSHV0 Suite', function () {
-    before(() => {
-    });
-
-    after(() => {
-    });
-
     utilsTests.run();
+
+    if (process.platform === 'win32') {
+        require('./L0WindowsUtilsTests').run();
+    } else {
+        require('./L0LinuxUtilsTests').run();
+    }
 });
